Create Audio Manager and implement in A note for testing
For handling the sounds an AudioManager is created, in which all the notes needed for the piano will be handled. For calling the sound when a piano key is pressed the PressKey script is created (this script should be modified in the future to add the movement when a key is pressed). Finally for testing the script is added to the current A note, using the A4 sound.