Storing data in local computer
Creation of AnswerData3D script inside classes folder, which describes an object with the variables to store in local computer.
Cration of DataManager script  inside scripts folder, which saves the user data inside the aplication root directory in UserData directory.
This is only for 3D level
AnswerManager is changed in order to  call for function in DataManager in charge of saving 3D lvl data