creating scene for user test
The TestLevel scene is created, this will be used to test initila user level. A new script TestAnswerManager is created in base of AnswerManager. For the moment it holds the same logic, this should be change to only test for a few intervals. Older scripts subscribed to AnswerManager Events are now also subscribe to TestAnswerManagerEvents