Modifying interval player to manage probabilities
The intervalPlayer script now has a variable to set whether or not to use dynamicProbability or not, and a dictionary to hold the probability of each interval used in the scene. A method is created to initialize the dictionary and called in start function. It does not have the logic for the dynamic probability yet (only has for default probability).
The interval definition is not yet afected by the dictionary